Registration Functionality
Before this, there was no authentication or protected routes so all the
users were accessing the same db but now that I've added a backend that
authenticates users, multiple users can sign up to the platform.
Also, once logged in the user will stay logged in because the jwt is
being fetched from local storage at every page reload.
Next steps: I believe the CRUD operations aren't working right now
so that needs to be fixed