CRUD finished
There were a lot of conflicts previously because the frontend and backend
were using entirely different schemas for tasks and that's why no CRUD
operation was working. I changed the schema and now everything is working
fine.
Next steps: Improve the UI for the toggle task feature and do something
about the speed of the app. It lags a lot because it fetches all the tasks
before rendering the page.