Zustand and Cookies
Implemented Zustand for state management. Was a REAL hassle but
the code is reduced somewhat and is easier to follow especially due
the cookies which really reduced the logic involved (since there's
no need to use localStorage)
For now, I've only changed the Register and Login routes to work with
them but this alone took 6 hours and is a major jump from what is
previously was so I'm making a separate commit out it.
Next Up: Email Verification functionality probably