Fixing vercel routing issues
I'm using react router, so when the user is navigated to a page via Link or
navigate, it was working properly. But as soon as I refresh the site while
being on one of these routes, Vercel was looking for a file by that route's name
which of course doesn't exist.